requirements.txt
python-dotenv
aerich
tortoise-orm[asyncpg]
python-telegram-bot
yt-dlp
sanic
sanic-ext
pydantic-settings
jinja2
